user/requirements.txt
Django==2.1.8
djangorestframework==3.9.2
djangorestframework-stubs==0.4.2
pytz==2018.9
psycopg2==2.7.7
django-reset-migrations==0.3.1
django-cors-middleware==1.3.1
django-oauth-toolkit==1.2.0
oauthlib==3.0.1
django-oauth2-provider==0.2.6.1
shortuuid==0.5.0
